refactor(#1116): split the declared-marshal comparator out of udt_canon.rs
`udt_canon.rs` was 1392 lines against the 800-line campsite limit, and #3935's
authority comments would have grown it further (the file-size ratchet FAILs any
change that enlarges an over-threshold file). Moving `CompareKind`,
`classify_comparator`, `comparator_supported_for`, `unsupported_comparator_err`,
`signed_value` and `compare_for_marshal` into a sibling `marshal_comparator.rs`
takes udt_canon to 1295 and gives the SECOND writer-side sorted-collection
comparator a file of its own, named for what it is.

Pure move: no behaviour change. `primitive_marshal_name` and the four moved items
are `pub(super)` (data_writer-internal), which is also what lets the #3935
differential test reach both comparators. Nothing outside udt_canon referenced any
of them before.

Also condenses this PR's read_assembly.rs comment fix to be line-neutral, since
that file is over threshold too (917 -> 917) and a split there is out of scope.
